Move JS views to own namespace
