+++ src/assets/javascripts/webpack-entry.jsx
